# Conflicts:
#	src/main/java/Views/UserView.java

+++ src/main/java/Views/UserView.java
@@ -5,20 +5,20 @@ import Models.Roles;
 import Models.User;
 
 import java.sql.SQLException;
+import javax.swing.text.View;
 import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class UserView {
-    // salvar coisas foreign como fkID ao invez de objeto
     private UserController controller = new UserController();
 
     public UserView() {
     }
 
     public void menu() {
-        Scanner input = new Scanner(System.in);
+        Scanner scanner = new Scanner(System.in);
         while (true) {
             System.out.println("-Menu Usuario-");
             System.out.println("1 - Crie um Usuario");
@@ -27,8 +27,11 @@ public class UserView {
             System.out.println("4 - Listar Usuarios");
             System.out.println("0 - Sair");
 
-            int choice = input.nextInt();
-            input.nextLine();
+            int choice = ViewUtils.getChoice(scanner);
+            if (choice == -1) {
+                System.out.println("Escolha invalida!");
+                continue;
+            }
 
             switch (choice) {
                 case 0:
@@ -54,18 +57,18 @@ public class UserView {
     }
 
     public void registrationMenu() {
-        Scanner input = new Scanner(System.in);
+        Scanner scanner = new Scanner(System.in);
         User user = new User();
 
         System.out.println("-Criar Usuario-");
         System.out.println("Escreva um nome:");
-        user.setName(input.nextLine());
+        user.setName(scanner.nextLine());
 
         System.out.println("Escreva uma senha:");
-        user.setNewPassword(input.nextLine());
+        user.setNewPassword(scanner.nextLine());
 
         System.out.println("Escreva um e-mail:");
-        user.setEmail(input.nextLine());
+        user.setEmail(scanner.nextLine());
 
         rolesMenu(user);
         try {
@@ -83,14 +86,17 @@ public class UserView {
                 return;
             }
 
-            Scanner input = new Scanner(System.in);
+            Scanner scanner = new Scanner(System.in);
 
-            System.out.println("-Editar Usuario-");
+            while (true) {System.out.println("-Editar Usuario-");
             list();
 
             System.out.println("Escolha um ID:");
-            Long id = input.nextLong();
-            input.nextLine();
+            Long id = (long) ViewUtils.getChoice(scanner);
+            if (id == -1) {
+            System.out.println("ID Invalido!");
+                continue;
+            }
 
             User choice = controller.getById(id);
             if (choice == null) {
@@ -99,42 +105,48 @@ public class UserView {
             }
 
             edit(choice);
-        }catch (SQLException e){
-            System.out.println("Falha ao Editar Usuario");
+            return;
         }
     }
 
     public void deleteMenu() {
-        Scanner input = new Scanner(System.in);
+        if (controller.getModels().size() == 0) {
+            System.out.println("Nao existe usuarios para deletar.");
+            return;
+        }
 
-        try {
-            if (controller.getModels().size() == 0) {
-                System.out.println("Nao existe usuarios para deletar.");
-                return;
-            }
+        Scanner scanner = new Scanner(System.in);
 
+        while (true) {
             System.out.println("-Deletar Usuario-");
             list();
 
             System.out.println("Escolha um ID:");
-            Long id = input.nextLong();
-            input.nextLine();
+            Long id = (long) ViewUtils.getChoice(scanner);
+            if (id == -1) {
+                System.out.println("ID Invalido!");
+                continue;
+            }
+
+            if (!login(controller.getById(id))) {
+                System.out.println("Senha incorreta!");
+                return;
+            }
 
             if (controller.deleteById(id)) {
                 System.out.println("Usuario deletado com sucesso!");
             } else {
                 System.out.println("Nao foi possivel deletar o Usuario, confirme se escreveu o ID correto.");
             }
-        }catch (SQLException e){
-            System.out.println("Falha ao Deletar Usuario");
+            return;
         }
     }
 
     public boolean login(User user) {
-        Scanner input = new Scanner(System.in);
+        Scanner scanner = new Scanner(System.in);
         System.out.println("Escreva sua senha:");
 
-        String password = input.nextLine();
+        String password = scanner.nextLine();
         return user.checkLogin(password);
     }
 
@@ -144,7 +156,7 @@ public class UserView {
             return;
         }
 
-        Scanner input = new Scanner(System.in);
+        Scanner scanner = new Scanner(System.in);
         while (true) {
             System.out.println("Informacoes:");
             System.out.printf("1 - %s: %s\n", "Nome", user.getName());
@@ -153,8 +165,11 @@ public class UserView {
             System.out.println("0 - Sair");
 
             System.out.println("Escolha uma opcao para mudar");
-            int choice = input.nextInt();
-            input.nextLine();
+            int choice = ViewUtils.getChoice(scanner);
+            if (choice == -1) {
+                System.out.println("Escolha invalida!");
+                continue;
+            }
 
             try {
                 switch (choice) {
@@ -162,12 +177,12 @@ public class UserView {
                         return;
                     case 1:
                         System.out.println("Escreva um novo nome:");
-                        user.setName(input.nextLine());
+                        user.setName(scanner.nextLine());
                         controller.editUser(user);
                         break;
                     case 2:
                         System.out.println("Escreva uma nova senha:");
-                        user.setNewPassword(input.nextLine());
+                        user.setNewPassword(scanner.nextLine());
                         controller.editUser(user);
                         break;
                     case 3:
@@ -195,13 +210,17 @@ public class UserView {
     }
 
     public void rolesMenu(User user) {
-        Scanner input = new Scanner(System.in);
+        Scanner scanner = new Scanner(System.in);
         while (true) {
             System.out.println("Cargos: " + user.getRoles());
             System.out.println("1 - Adicionar Cargo");
             System.out.println("2 - Remover Cargo");
             System.out.println("3 - Confirmar");
-            int choice = input.nextInt();
+            int choice = ViewUtils.getChoice(scanner);
+            if (choice == -1) {
+                System.out.println("Escolha invalida!");
+                continue;
+            }
 
             if (choice == 3) {
                 return;
@@ -219,9 +238,13 @@ public class UserView {
                 System.out.printf("%d - %s\n", i+1, possibleRoles.get(i));
             }
 
-            int selectedRole = input.nextInt() - 1;
-            input.nextLine();
+            int selectedRole = ViewUtils.getChoice(scanner);
+            if (selectedRole == -1) {
+                System.out.println("Escolha invalida!");
+                continue;
+            }
 
+            selectedRole -= 1;
             if (selectedRole >= 0 && selectedRole < possibleRoles.size()) {
                 Roles role = Roles.valueOf(possibleRoles.get(selectedRole));
